Move gql model in domain & scaling group
