+++ src/ai/backend/manager/models/domain.py
@@ -6,8 +6,6 @@ from dataclasses import dataclass, field
 from datetime import datetime
 from typing import (
     TYPE_CHECKING,
-    Any,
-    Callable,
     List,
     Optional,
     Self,
@@ -18,22 +16,15 @@ from typing import (
     override,
 )
 
-import graphene
 import sqlalchemy as sa
-from graphene.types.datetime import DateTime as GQLDateTime
-from graphql import Undefined
 from sqlalchemy.dialects import postgresql as pgsql
-from sqlalchemy.engine.row import Row
 from sqlalchemy.ext.asyncio import AsyncConnection as SAConnection
 from sqlalchemy.ext.asyncio import AsyncSession as SASession
 from sqlalchemy.orm import load_only, relationship
 
 from ai.backend.common import msgpack
-from ai.backend.common.types import ResourceSlot, Sentinel, VFolderHostPermissionMap
+from ai.backend.common.types import VFolderHostPermissionMap
 from ai.backend.logging import BraceStyleAdapter
-from ai.backend.manager.models.utils import define_state
-from ai.backend.manager.services.domain.types import DomainCreator, DomainModifier
-from ai.backend.manager.types import OptionalState, TriState
 
 from ..defs import RESERVED_DOTFILES
 from .base import (
@@ -41,7 +32,6 @@ from .base import (
     ResourceSlotColumn,
     SlugType,
     VFolderHostPermissionColumn,
-    batch_result,
     mapper_registry,
 )
 from .rbac import (
@@ -57,29 +47,9 @@ from .rbac import (
 )
 from .rbac.context import ClientContext
 from .rbac.permission_defs import DomainPermission
-from .scaling_group import ScalingGroup
-from .user import UserRole
 
 if TYPE_CHECKING:
-    from ai.backend.manager.services.domain.actions.create_domain import (
-        CreateDomainAction,
-        CreateDomainActionResult,
-    )
-    from ai.backend.manager.services.domain.actions.delete_domain import (
-        DeleteDomainAction,
-        DeleteDomainActionResult,
-    )
-    from ai.backend.manager.services.domain.actions.modify_domain import (
-        ModifyDomainAction,
-        ModifyDomainActionResult,
-    )
-    from ai.backend.manager.services.domain.actions.purge_domain import (
-        PurgeDomainAction,
-        PurgeDomainActionResult,
-    )
-    from ai.backend.manager.services.domain.types import DomainData
-
-    from .gql import GraphQueryContext
+    pass
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))
 
@@ -87,12 +57,6 @@ log = BraceStyleAdapter(logging.getLogger(__spec__.name))
 __all__: Sequence[str] = (
     "domains",
     "DomainRow",
-    "Domain",
-    "DomainInput",
-    "ModifyDomainInput",
-    "CreateDomain",
-    "ModifyDomain",
-    "DeleteDomain",
     "DomainDotfile",
     "MAXIMUM_DOTFILE_SIZE",
     "query_domain_dotfiles",
@@ -210,309 +174,6 @@ class DomainModel(RBACModel[DomainPermission]):
             _permissions=frozenset(permissions),
             orm_obj=row,
         )
-
-
-class Domain(graphene.ObjectType):
-    name = graphene.String()
-    description = graphene.String()
-    is_active = graphene.Boolean()
-    created_at = GQLDateTime()
-    modified_at = GQLDateTime()
-    total_resource_slots = graphene.JSONString()
-    allowed_vfolder_hosts = graphene.JSONString()
-    allowed_docker_registries = graphene.List(lambda: graphene.String)
-    integration_id = graphene.String()
-
-    # Dynamic fields.
-    scaling_groups = graphene.List(lambda: graphene.String)
-
-    async def resolve_scaling_groups(self, info: graphene.ResolveInfo) -> Sequence[str]:
-        sgroups = await ScalingGroup.load_by_domain(info.context, self.name)
-        return [sg.name for sg in sgroups]
-
-    @classmethod
-    def from_row(cls, ctx: GraphQueryContext, row: Row) -> Optional[Domain]:
-        if row is None:
-            return None
-        return cls(
-            name=row["name"],
-            description=row["description"],
-            is_active=row["is_active"],
-            created_at=row["created_at"],
-            modified_at=row["modified_at"],
-            total_resource_slots=(
-                row["total_resource_slots"].to_json()
-                if row["total_resource_slots"] is not None
-                else {}
-            ),
-            allowed_vfolder_hosts=row["allowed_vfolder_hosts"].to_json(),
-            allowed_docker_registries=row["allowed_docker_registries"],
-            integration_id=row["integration_id"],
-        )
-
-    @classmethod
-    def from_dto(cls, dto: DomainData) -> Domain:
-        return cls(
-            name=dto.name,
-            description=dto.description,
-            is_active=dto.is_active,
-            created_at=dto.created_at,
-            modified_at=dto.modified_at,
-            total_resource_slots=dto.total_resource_slots.to_json()
-            if dto.total_resource_slots
-            else {},
-            allowed_vfolder_hosts=dto.allowed_vfolder_hosts.to_json(),
-            allowed_docker_registries=dto.allowed_docker_registries,
-            integration_id=dto.integration_id,
-        )
-
-    @classmethod
-    async def load_all(
-        cls,
-        ctx: GraphQueryContext,
-        *,
-        is_active: Optional[bool] = None,
-    ) -> Sequence[Domain]:
-        async with ctx.db.begin_readonly() as conn:
-            query = sa.select([domains]).select_from(domains)
-            if is_active is not None:
-                query = query.where(domains.c.is_active == is_active)
-            return [
-                obj
-                async for row in (await conn.stream(query))
-                if (obj := cls.from_row(ctx, row)) is not None
-            ]
-
-    @classmethod
-    async def batch_load_by_name(
-        cls,
-        ctx: GraphQueryContext,
-        names: Sequence[str],
-        *,
-        is_active: Optional[bool] = None,
-    ) -> Sequence[Optional[Domain]]:
-        async with ctx.db.begin_readonly() as conn:
-            query = sa.select([domains]).select_from(domains).where(domains.c.name.in_(names))
-            if is_active is not None:
-                query = query.where(domains.c.is_active == is_active)
-            return await batch_result(
-                ctx,
-                conn,
-                query,
-                cls,
-                names,
-                lambda row: row["name"],
-            )
-
-
-class DomainInput(graphene.InputObjectType):
-    description = graphene.String(required=False, default_value="")
-    is_active = graphene.Boolean(required=False, default_value=True)
-    total_resource_slots = graphene.JSONString(required=False, default_value={})
-    allowed_vfolder_hosts = graphene.JSONString(required=False, default_value={})
-    allowed_docker_registries = graphene.List(
-        lambda: graphene.String, required=False, default_value=[]
-    )
-    integration_id = graphene.String(required=False, default_value=None)
-
-    def to_action(self, domain_name: str) -> CreateDomainAction:
-        def value_or_none(value):
-            return value if value is not Undefined else None
-
-        return CreateDomainAction(
-            input=DomainCreator(
-                name=domain_name,
-                description=value_or_none(self.description),
-                is_active=value_or_none(self.is_active),
-                total_resource_slots=value_or_none(self.total_resource_slots),
-                allowed_vfolder_hosts=value_or_none(self.allowed_vfolder_hosts),
-                allowed_docker_registries=value_or_none(self.allowed_docker_registries),
-                integration_id=value_or_none(self.integration_id),
-            ),
-        )
-
-
-class ModifyDomainInput(graphene.InputObjectType):
-    name = graphene.String(required=False)
-    description = graphene.String(required=False)
-    is_active = graphene.Boolean(required=False)
-    total_resource_slots = graphene.JSONString(required=False)
-    allowed_vfolder_hosts = graphene.JSONString(required=False)
-    allowed_docker_registries = graphene.List(lambda: graphene.String, required=False)
-    integration_id = graphene.String(required=False)
-
-    def _convert_field(
-        self, field_value: Any, converter: Optional[Callable[[Any], Any]] = None
-    ) -> Any | Sentinel:
-        if field_value is Undefined:
-            return Sentinel.TOKEN
-        if converter is not None:
-            return converter(field_value)
-        return field_value
-
-    def to_action(self, domain_name: str) -> ModifyDomainAction:
-        def value_or_none(value):
-            return value if value is not Undefined else None
-
-        return ModifyDomainAction(
-            domain_name=domain_name,
-            modifier=DomainModifier(
-                name=OptionalState("name", define_state(self.name), value_or_none(self.name)),
-                description=TriState(
-                    "description",
-                    define_state(self.description),
-                    value_or_none(self.description),
-                ),
-                is_active=OptionalState(
-                    "is_active", define_state(self.is_active), value_or_none(self.is_active)
-                ),
-                total_resource_slots=TriState(
-                    "total_resource_slots",
-                    define_state(self.total_resource_slots),
-                    None
-                    if self.total_resource_slots is Undefined
-                    else ResourceSlot.from_user_input(self.total_resource_slots, None),
-                ),
-                allowed_vfolder_hosts=OptionalState(
-                    "allowed_vfolder_hosts",
-                    define_state(self.allowed_vfolder_hosts),
-                    value_or_none(self.allowed_vfolder_hosts),
-                ),
-                allowed_docker_registries=OptionalState(
-                    "allowed_docker_registries",
-                    define_state(self.allowed_docker_registries),
-                    value_or_none(self.allowed_vfolder_hosts),
-                ),
-                integration_id=TriState(
-                    "integration_id",
-                    define_state(self.integration_id),
-                    value_or_none(self.integration_id),
-                ),
-            ),
-        )
-
-
-class CreateDomain(graphene.Mutation):
-    allowed_roles = (UserRole.SUPERADMIN,)
-
-    class Arguments:
-        name = graphene.String(required=True)
-        props = DomainInput(required=True)
-
-    ok = graphene.Boolean()
-    msg = graphene.String()
-    domain = graphene.Field(lambda: Domain, required=False)
-
-    @classmethod
-    async def mutate(
-        cls,
-        root,
-        info: graphene.ResolveInfo,
-        name: str,
-        props: DomainInput,
-    ) -> CreateDomain:
-        ctx: GraphQueryContext = info.context
-
-        action: CreateDomainAction = props.to_action(name)
-        res: CreateDomainActionResult = await ctx.processors.domain.create_domain.wait_for_complete(
-            action
-        )
-
-        domain_data: Optional[DomainData] = res.domain_data
-
-        return cls(
-            ok=res.success,
-            msg=res.description,
-            domain=Domain.from_dto(domain_data) if domain_data else None,
-        )
-
-
-class ModifyDomain(graphene.Mutation):
-    allowed_roles = (UserRole.SUPERADMIN,)
-
-    class Arguments:
-        name = graphene.String(required=True)
-        props = ModifyDomainInput(required=True)
-
-    ok = graphene.Boolean()
-    msg = graphene.String()
-    domain = graphene.Field(lambda: Domain, required=False)
-
-    @classmethod
-    async def mutate(
-        cls,
-        root,
-        info: graphene.ResolveInfo,
-        name: str,
-        props: ModifyDomainInput,
-    ) -> ModifyDomain:
-        ctx: GraphQueryContext = info.context
-
-        action: ModifyDomainAction = props.to_action(name)
-        res: ModifyDomainActionResult = await ctx.processors.domain.modify_domain.wait_for_complete(
-            action
-        )
-
-        domain_data: Optional[DomainData] = res.domain_data
-
-        return cls(
-            ok=res.success,
-            msg=res.description,
-            domain=Domain.from_dto(domain_data) if domain_data else None,
-        )
-
-
-class DeleteDomain(graphene.Mutation):
-    """
-    Instead of deleting the domain, just mark it as inactive.
-    """
-
-    allowed_roles = (UserRole.SUPERADMIN,)
-
-    class Arguments:
-        name = graphene.String(required=True)
-
-    ok = graphene.Boolean()
-    msg = graphene.String()
-
-    @classmethod
-    async def mutate(cls, root, info: graphene.ResolveInfo, name: str) -> DeleteDomain:
-        ctx: GraphQueryContext = info.context
-
-        action = DeleteDomainAction(name)
-        res: DeleteDomainActionResult = await ctx.processors.domain.delete_domain.wait_for_complete(
-            action
-        )
-
-        return cls(ok=res.success, msg=res.description)
-
-
-class PurgeDomain(graphene.Mutation):
-    """
-    Completely delete domain from DB.
-
-    Domain-bound kernels will also be all deleted.
-    To purge domain, there should be no users and groups in the target domain.
-    """
-
-    allowed_roles = (UserRole.SUPERADMIN,)
-
-    class Arguments:
-        name = graphene.String(required=True)
-
-    ok = graphene.Boolean()
-    msg = graphene.String()
-
-    @classmethod
-    async def mutate(cls, root, info: graphene.ResolveInfo, name: str) -> PurgeDomain:
-        ctx: GraphQueryContext = info.context
-
-        action = PurgeDomainAction(name)
-        res: PurgeDomainActionResult = await ctx.processors.domain.purge_domain.wait_for_complete(
-            action
-        )
-
-        return cls(ok=res.success, msg=res.description)
 
 
 class DomainDotfile(TypedDict):
